move exceptions file
